提交

+++ Tzg-Biz/src/main/java/com/tzg/entitys/kff/project/KFFProject.java
@@ -56,7 +56,7 @@ public class KFFProject implements Serializable {
 	/**
 	 * 发行数量
 	 */
-	private java.lang.Integer issueNum = 0;
+	private java.lang.Long issueNum = 0L;
 	/**
 	 * 白皮书地址
 	 */
@@ -229,11 +229,11 @@ public class KFFProject implements Serializable {
 		return this.issueDateStr;
 	}
 
-	public void setIssueNum(java.lang.Integer value) {
+	public void setIssueNum(java.lang.Long value) {
 		this.issueNum = value;
 	}
 
-	public java.lang.Integer getIssueNum() {
+	public java.lang.Long getIssueNum() {
 		return this.issueNum;
 	}
